feat(live-activities): expose as SDK-managed module via CustomerIO.liveActivities
Register Live Activities with SDKConfigBuilder.addModule(LiveActivitiesModule(config:))
so it initializes during CustomerIO.initialize(withConfig:), and reach it through
CustomerIO.liveActivities — mirroring the Location module pattern. The public API moves to
a LiveActivitiesInstance protocol backed by a shared state holder plus an uninitialized
stub; the stub logs and no-ops (start/adopt return nil) so an initialization-timing race
never throws in a host app. Removes the host-held static initialize entry point and the
superseded handleDeepLinkOpen.
