+++ Apps/APN-UIKit/APN UIKit/SceneDelegate.swift
@@ -1,3 +1,4 @@
+import CioDataPipelines
 import CioLiveActivities
 import UIKit
 
@@ -74,11 +75,11 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     private func handle(urlContexts: Set<UIOpenURLContext>) {
         for context in urlContexts {
             var url = context.url
-            if #available(iOS 17.2, *), let liveActivities = AppDelegate.liveActivities {
+            if #available(iOS 17.2, *) {
                 // A tap on a Customer.io Live Activity arrives as a CIO tracking URL: report the
                 // `opened` for the exact delivery shown and unwrap the customer's deep link (nil if
                 // none). Non-CIO URLs are returned unchanged.
-                guard let destination = liveActivities.handleWidgetUrl(url) else { continue }
+                guard let destination = CustomerIO.liveActivities.handleWidgetUrl(url) else { continue }
                 url = destination
                 if url.host == LiveActivitiesViewController.deepLinkHost {
                     routeToLiveActivities()
