+++ Sources/LiveActivities/Config/LiveActivityConfig.swift
@@ -4,14 +4,15 @@ import Foundation
 /// Configuration for the Live Activities module.
 ///
 /// Build instances via `LiveActivityConfigBuilder` and pass the result to
-/// `LiveActivitiesModule.initialize(_:)`:
+/// `LiveActivitiesModule(config:)`, then register it before `CustomerIO.initialize(withConfig:)`:
 /// ```swift
+/// let config = SDKConfigBuilder(cdpApiKey: "your_key")
+///     .addModule(LiveActivitiesModule(config:
+///         LiveActivityConfigBuilder()
+///             .register(OrderAttributes.self, identifier: "io.customer.livenotifications.order")
+///             .build()))
+///     .build()
 /// CustomerIO.initialize(withConfig: config)
-/// LiveActivitiesModule.initialize(
-///     LiveActivityConfigBuilder()
-///         .register(OrderAttributes.self, identifier: "io.customer.livenotifications.order")
-///         .build()
-/// )
 /// ```
 public struct LiveActivityConfig {
     /// Module-level log level override. When `nil`, the SDK-wide log level is used.
