+++ Sources/LiveActivities/LiveActivitiesModule.swift
@@ -1,368 +1,38 @@
 import CioInternalCommon
-import CioLiveActivities_Attributes
 import Foundation
 
-#if os(iOS)
-import ActivityKit
-#endif
-
-/// Errors thrown by the Live Activities public API.
-public enum LiveActivityError: Error {
-    /// `start` was called for an attributes type that was not passed to
-    /// `LiveActivityConfigBuilder.register(_:identifier:)`.
-    case typeNotRegistered(String)
-}
-
-/// Live Activities module for the Customer.io SDK.
+/// Live Activities module. Register it via `SDKConfigBuilder.addModule(_:)` so it is initialized
+/// during `CustomerIO.initialize(withConfig:)`, then access it through ``CustomerIO/liveActivities``.
 ///
-/// Call `initialize` after `CustomerIO.initialize(withConfig:)` and hold the returned instance
-/// for the lifetime of your app:
+/// **Example:**
 /// ```swift
+/// let config = SDKConfigBuilder(cdpApiKey: "your_key")
+///     .addModule(LiveActivitiesModule(config:
+///         LiveActivityConfigBuilder()
+///             .register(CIOSegmentsAttributes.self)
+///             .register(CIOCountdownTimerAttributes.self)
+///             .build()))
+///     .build()
 /// CustomerIO.initialize(withConfig: config)
-/// let liveActivities = LiveActivitiesModule.initialize(
-///     LiveActivityConfigBuilder()
-///         .register(OrderAttributes.self, identifier: "io.customer.livenotifications.order")
-///         .build()
-/// )
-/// // Start an activity locally — the SDK mints its id and reports a `start` event:
-/// let handle = try liveActivities.start(contentState: .init(...)) { id in
-///     OrderAttributes(cioInstanceId: id, ...)
-/// }
-/// await handle.update(.init(...))
-/// await handle.end(.init(...))
 /// ```
-public final class LiveActivitiesModule {
+public final class LiveActivitiesModule: CustomerIOModule {
+    public let moduleName: String = "LiveActivities"
     private let config: LiveActivityConfig
-    private let sdk: CIOLiveActivitiesSDKProviding
-    private let tokenStorage: LiveActivityTokenStorage
 
-    private let identity = LiveActivityIdentity()
-    private let localEndTracker: LiveActivityLocalEndTracker
-    private let reporter: LiveActivityReporter
-    private let registrar: LiveActivityRegistrar
-    private let deliveryTracker: LiveActivityDeliveryTracker
-    private let observer: LiveActivityObserver
-
-    /// Latest delivery metadata observed per activity instance, used to attribute an `opened`
-    /// metric and resolve the deep link when the app is opened from a tapped activity.
-    private let latestMetadata = Synchronized<[String: CIOLiveActivityMetadata]>([:])
-
-    // MARK: - Entry point
-
-    /// Initialize the Live Activities module. Call after `CustomerIO.initialize(withConfig:)`.
-    /// Hold the returned instance for the app's lifetime — it is not a singleton.
-    @discardableResult
-    public static func initialize(_ config: LiveActivityConfig) -> LiveActivitiesModule {
-        let sdk = CustomerIO.shared
-        let module = LiveActivitiesModule(
-            config: config,
-            sdk: sdk,
-            tokenStorage: KeyValueLiveActivityTokenStore(storage: sdk.sharedKeyValueStorage)
-        )
-        module.performInitialization()
-        return module
-    }
-
-    // MARK: - Init (also used by tests)
-
-    init(
-        config: LiveActivityConfig,
-        sdk: CIOLiveActivitiesSDKProviding,
-        tokenStorage: LiveActivityTokenStorage,
-        localEndTracker: LiveActivityLocalEndTracker = LiveActivityLocalEndTracker()
-    ) {
+    public init(config: LiveActivityConfig) {
         self.config = config
-        self.sdk = sdk
-        self.tokenStorage = tokenStorage
-        self.localEndTracker = localEndTracker
-
-        let identity = self.identity
-        let reporter = LiveActivityReporter(
-            track: { name, properties in sdk.track(name: name, properties: properties) },
-            // Gate on the SDK's synchronous identified flag (backed by DataPipelineTracking),
-            // not the async event-bus-fed identity cache, so a start() right after identify()
-            // isn't dropped. Token registration still uses `identity` via the registrar.
-            isUserIdentified: { sdk.isUserIdentified },
-            deviceToken: { identity.deviceToken },
-            logger: sdk.logger
-        )
-        let registrar = LiveActivityRegistrar(identity: identity, store: tokenStorage, reporter: reporter)
-        self.reporter = reporter
-        self.registrar = registrar
-
-        let deliveryTracker = LiveActivityDeliveryTracker(
-            postMetric: { [sdk] deliveryId, event, deliveryToken in
-                sdk.eventBusHandler.postEvent(
-                    TrackMetricEvent(deliveryID: deliveryId, event: event, deviceToken: deliveryToken)
-                )
-            },
-            store: tokenStorage,
-            logger: sdk.logger
-        )
-        self.deliveryTracker = deliveryTracker
-
-        self.observer = LiveActivityObserver(
-            registrations: config.registrations,
-            registrar: registrar,
-            localEndTracker: localEndTracker,
-            store: tokenStorage,
-            onUserDismissed: { [reporter] notificationType, cioInstanceId in
-                // A user swipe-away — report `end` (matches Android's dismiss-intent behavior).
-                reporter.reportEnd(instanceUUID: cioInstanceId, notificationType: notificationType)
-            },
-            onContentMetadata: { [deliveryTracker, latestMetadata] _, cioInstanceId, metadata in
-                // A backend push (start or update) carrying Customer.io delivery metadata: report a
-                // `delivered` receipt (deduped by delivery id) and remember the tap destination.
-                latestMetadata.mutating { $0[cioInstanceId] = metadata }
-                deliveryTracker.reportDelivered(metadata: metadata)
-            },
-            onActivityEnded: { [latestMetadata] cioInstanceId in
-                // Drop the tap destination once the activity terminates, so a later open of the same
-                // URL can't be misattributed to this (ended) delivery.
-                latestMetadata.mutating { $0[cioInstanceId] = nil }
-            }
-        )
     }
 
-    // MARK: - Local lifecycle API
-
-    #if os(iOS)
-    /// Start a Live Activity locally. Pass your fully-built `attributes` and initial
-    /// `contentState`; the SDK requests the activity, mints and tracks its correlation id
-    /// internally, and reports a `start` event. You never supply or see the id at the call site.
-    ///
-    /// Works with any `ActivityAttributes` type — conforming to `CIOActivityAttribute` is only
-    /// required for push-to-start.
-    ///
-    /// - Returns: A handle whose `end` reports an `end` event; `update` applies the new
-    ///   content-state locally and is intentionally not reported.
-    /// - Throws: `LiveActivityError.typeNotRegistered` if `Attributes` was not registered.
-    @available(iOS 16.2, *)
-    @discardableResult
-    public func start<Attributes: ActivityAttributes>(
-        _ attributes: Attributes,
-        contentState: Attributes.ContentState,
-        staleDate: Date? = nil,
-        relevanceScore: Double = 0
-    ) throws -> CIOLiveActivity<Attributes> {
-        guard let notificationType = notificationType(forTypeName: String(describing: Attributes.self)) else {
-            throw LiveActivityError.typeNotRegistered(String(describing: Attributes.self))
-        }
-        let activity = try Activity.request(
-            attributes: attributes,
-            content: ActivityContent(state: contentState, staleDate: staleDate, relevanceScore: relevanceScore),
-            pushType: .token
-        )
-        // Mint (and persist) the correlation id keyed by the system Activity.id, atomically so the
-        // registration observer picking up this same activity resolves to the identical id.
-        let id = tokenStorage.resolveInstanceId(forActivityId: activity.id) { ULID.generate() }
-        reporter.reportStart(
-            instanceUUID: id,
-            notificationType: notificationType,
-            attributes: LiveActivityReporter.encode(attributes),
-            contentState: LiveActivityReporter.encode(contentState)
-        )
-        return CIOLiveActivity(
-            id: id,
-            activity: activity,
-            reporter: reporter,
-            notificationType: notificationType,
-            logger: sdk.logger,
-            markLocalEnd: { [localEndTracker] in localEndTracker.markEnded($0) }
-        )
-    }
-
-    /// Start a Live Activity locally for a `CIOActivityAttribute` type. Same as the generic
-    /// `start`, except the SDK mints the correlation id and **injects it into `cioInstanceId`** on
-    /// your attributes *before* requesting the activity — so the live activity itself, the reported
-    /// `start` event, and later token registrations all carry the identical id, matching the
-    /// push-to-start contract. Overload resolution prefers this over the generic `start` whenever
-    /// `Attributes` conforms to `CIOActivityAttribute`.
-    @available(iOS 16.2, *)
-    @discardableResult
-    public func start<Attributes: CIOActivityAttribute>(
-        _ attributes: Attributes,
-        contentState: Attributes.ContentState,
-        staleDate: Date? = nil,
-        relevanceScore: Double = 0
-    ) throws -> CIOLiveActivity<Attributes> {
-        guard let notificationType = notificationType(forTypeName: String(describing: Attributes.self)) else {
-            throw LiveActivityError.typeNotRegistered(String(describing: Attributes.self))
-        }
-        // Mint first and inject, so the activity carries the id (a plain ActivityAttributes type has
-        // no field to hold it). Persist the Activity.id → id mapping so the observer and relaunch
-        // recovery resolve the identical id even though the mint happened here.
-        let id = ULID.generate()
-        var attributes = attributes
-        attributes.cioInstanceId = id
-        let activity = try Activity.request(
-            attributes: attributes,
-            content: ActivityContent(state: contentState, staleDate: staleDate, relevanceScore: relevanceScore),
-            pushType: .token
-        )
-        _ = tokenStorage.resolveInstanceId(forActivityId: activity.id) { id }
-        reporter.reportStart(
-            instanceUUID: id,
-            notificationType: notificationType,
-            attributes: LiveActivityReporter.encode(attributes),
-            contentState: LiveActivityReporter.encode(contentState)
-        )
-        return CIOLiveActivity(
-            id: id,
-            activity: activity,
-            reporter: reporter,
-            notificationType: notificationType,
-            logger: sdk.logger,
-            markLocalEnd: { [localEndTracker] in localEndTracker.markEnded($0) }
-        )
-    }
-
-    /// Wrap an activity your app created directly, so you can drive `update`/`end` through the
-    /// returned handle — only `end` is reported; `update` applies locally and is not reported. Does
-    /// not report a `start` event (use `start` for that). Token capture for
-    /// registered types happens automatically via observation regardless of `adopt`. Works with any
-    /// `ActivityAttributes` type.
-    @available(iOS 16.2, *)
-    @discardableResult
-    public func adopt<Attributes: ActivityAttributes>(_ activity: Activity<Attributes>) -> CIOLiveActivity<Attributes> {
-        let notificationType = notificationType(forTypeName: String(describing: Attributes.self))
-            ?? String(describing: Attributes.self)
-        let id = tokenStorage.resolveInstanceId(forActivityId: activity.id) { ULID.generate() }
-        return CIOLiveActivity(
-            id: id,
-            activity: activity,
-            reporter: reporter,
-            notificationType: notificationType,
-            logger: sdk.logger,
-            markLocalEnd: { [localEndTracker] in localEndTracker.markEnded($0) }
-        )
-    }
-    #endif
-
-    // MARK: - Deep link / open tracking
-
-    /// Report an `opened` metric when the app is opened from a tapped Live Activity. Call this from
-    /// your `UIApplicationDelegate.application(_:open:options:)` /
-    /// `UISceneDelegate.scene(_:openURLContexts:)` — the same place you already route your app's
-    /// deep links — mirroring how push taps are forwarded to the SDK.
-    ///
-    /// If the URL matches a Customer.io-tracked activity's deep link, the SDK reports an `opened`
-    /// metric (the same metric normal push uses) and returns `true`. Returns `false` for unrelated
-    /// URLs. Navigation stays with your existing URL routing: because a Live Activity tap arrives
-    /// through the normal `openURL` path (not a push-tap callback), the SDK does not re-open the URL
-    /// — that would re-enter `openURL` and loop.
-    @discardableResult
-    public func handleDeepLinkOpen(_ url: URL) -> Bool {
-        let target = url.absoluteString
-        // Match without consuming: `opened` is intentionally NOT deduped — each tap is a distinct
-        // open (matching push). Staleness is handled by evicting metadata when the activity ends
-        // (see the observer's `onActivityEnded`), not here. Only a push-backed activity (non-empty
-        // deliveryId) yields an `opened` metric, so require one — otherwise return false rather than
-        // claim we handled an open we didn't report (e.g. a locally-started activity with no id).
-        guard let metadata = latestMetadata.wrappedValue.values.first(where: {
-            $0.deepLink == target && !($0.deliveryId ?? "").isEmpty
-        }) else { return false }
-        deliveryTracker.reportOpened(metadata: metadata)
-        return true
-    }
-
-    /// Report an `opened` for a Customer.io Live Activity tracking URL (the one the `cioWidgetUrl(_:)`
-    /// widget modifier builds), and return the customer's deep link to navigate to.
-    ///
-    /// Call this from your `UISceneDelegate.scene(_:openURLContexts:)` /
-    /// `UIApplicationDelegate.application(_:open:options:)` for each opened URL:
-    /// - If `url` is a Customer.io tracking URL, the SDK reports an `opened` for the **exact delivery**
-    ///   the user tapped — the identity is carried in the URL, so it needs no observer lookup and works
-    ///   even from a cold launch — and returns its `redirect` deep link (`nil` if none) for you to route.
-    /// - Otherwise `url` is returned unchanged, so your existing routing handles non-CIO URLs.
-    @discardableResult
-    public func handleWidgetUrl(_ url: URL) -> URL? {
-        guard let parsed = CioLiveActivityWidgetUrl.parse(url) else { return url }
-        if let deliveryId = parsed.deliveryId, !deliveryId.isEmpty {
-            deliveryTracker.reportOpened(
-                metadata: CIOLiveActivityMetadata(deliveryId: deliveryId, deliveryToken: parsed.deliveryToken)
-            )
-        }
-        return parsed.redirect
-    }
-
-    // MARK: - Private
-
-    private func notificationType(forTypeName name: String) -> String? {
-        config.registrations.first { $0.attributesTypeName == name }?.activityIdentifier
-    }
-
-    private func performInitialization() {
-        // Apply the optional module log-level override. The SDK exposes a single shared logger, so
-        // this raises/lowers the level used by Live Activities logging (and the shared logger with it);
-        // when `nil` the existing SDK-wide level is left untouched.
-        if let logLevel = config.logLevel {
-            sdk.logger.setLogLevel(logLevel)
-        }
-
-        sdk.logger.debug("LiveActivities module initialized.", "LiveActivities")
-
-        identity.deviceToken = sdk.registeredDeviceToken
-
-        // Seed any persisted push-to-start token before adding observers: the event-bus replays the
-        // last identify / device-token events to the new observers, and that replay flushes the
-        // seeded token — so registration no longer depends on ActivityKit re-yielding the token.
-        registrar.seedPendingPushToStartFromStore()
-
-        registerEventBusObservers()
-        observer.start()
-    }
-
-    private func registerEventBusObservers() {
-        sdk.eventBusHandler.addObserver(ProfileIdentifiedEvent.self) { [identity, registrar] event in
-            identity.userId = event.identifier
-            registrar.reevaluate()
-        }
-        sdk.eventBusHandler.addObserver(AnonymousProfileIdentifiedEvent.self) { [identity] _ in
-            // An anonymous profile is not an identified user — Live Activities stay gated off.
-            identity.userId = nil
-        }
-        sdk.eventBusHandler.addObserver(RegisterDeviceTokenEvent.self) { [identity, registrar] event in
-            identity.deviceToken = event.token
-            registrar.reevaluate()
-        }
-        sdk.eventBusHandler.addObserver(DeleteDeviceTokenEvent.self) { [identity] _ in
-            identity.deviceToken = nil
-        }
-        sdk.eventBusHandler.addObserver(ResetEvent.self) { [weak self] _ in
-            Task { [weak self] in await self?.handleReset() }
-        }
-    }
-
-    // Internal (not private) so the reset-suppression behavior can be driven directly in tests
-    // without routing through the async event bus.
-    func handleReset() async {
-        // Reset is a logout: force-end every activity so one user's Live Activities never linger
-        // into the next session. No `end` event is reported for these — the user is being
-        // de-identified, so lifecycle events must not fire.
+    public func initialize() {
+        // No main-thread requirement (unlike Location's CLLocationManager): registering event-bus
+        // observers and starting ActivityKit observation are thread-agnostic, so we initialize on
+        // whatever thread CustomerIO.initialize(withConfig:) ran on — a wrapper initializing off the
+        // main thread must still get a working module, not a silently-disabled stub.
         //
-        // Each activity is marked as an SDK-initiated (local) end just before it is force-ended, so
-        // the `.dismissed` terminal its immediate dismissal produces is consumed by the observer and
-        // not reported as a user swipe. Identity-based gating can't suppress this on its own: an
-        // account switch — clearIdentify() then identify(B) — puts B into the synchronous identity
-        // store before this async reset runs, so A's forced ends would otherwise be reported under B.
-        identity.userId = nil
-        // Do NOT clear existing local-end markers here: an activity the app ended via
-        // CIOLiveActivity.end() just before logout is already gone from Activity.activities, so the
-        // force-end loop below won't re-mark it. Dropping its marker would let its late `.dismissed`
-        // terminal be misreported as a user swipe (under the next user). Markers are in-memory,
-        // one-shot, and ULID-keyed, so keeping them is inert for any other activity.
-        latestMetadata.wrappedValue = [:]
-        for registration in config.registrations {
-            await registration.endAllActivities { [tokenStorage, localEndTracker] activityId, attributesInstanceId in
-                let instanceId = tokenStorage.resolveInstanceId(forActivityId: activityId) {
-                    if let attributesInstanceId, !attributesInstanceId.isEmpty { return attributesInstanceId }
-                    return ULID.generate()
-                }
-                localEndTracker.markEnded(instanceId)
-            }
+        // Live Activities require iOS 16.2+. On older systems the module stays uninitialized (its
+        // accessor returns the no-op stub); nothing to observe or seed there.
+        if #available(iOS 16.2, *) {
+            LiveActivitiesModuleState.shared.performInitialization(config: config)
         }
-        registrar.handleReset()
-        observer.restart()
     }
 }
